+++ astroquery/mast/__init__.py
@@ -31,13 +31,13 @@ class Conf(_config.ConfigNamespace):
 conf = Conf()
 
 from .cutouts import TesscutClass, Tesscut, ZcutClass, Zcut
-from .observations import Observations, ObservationsClass, MastClass, Mast
+from .observations import MastObservations, MastObservationsClass, MastClass, Mast
 from .collections import Catalogs, CatalogsClass
 from .missions import MastMissions, MastMissionsClass
 from .core import MastQueryWithLogin
 from . import utils
 
-__all__ = ['Observations', 'ObservationsClass',
+__all__ = ['MastObservations', 'MastObservationsClass',
            'Catalogs', 'CatalogsClass',
            'MastMissions', 'MastMissionsClass',
            'Mast', 'MastClass',